wip: ban user

+++ instantTchat/src/stores/channel-store.js
@@ -106,24 +106,29 @@ const useChannelStore = defineStore({
               }
           }
         },
-        async removeUserFromChannel(token, id, params) {
-          try{
-            const response = await api.delete(`/protected/channel/${id}/user/${user.id}`, params, {
-              headers: {
-                'Authorization': `Bearer ${token}`
-              }
-            })
-            this.channels = this.channels.map(channel => {
-              if(channel.id === id) {
-                return response.data
-              } else {
-                return channel
-              }
-            })
-            return response.data
-          }
-          catch(error) {
-            throw error
+        async removeUserFromChannel(token, id, params, channelCreator) {
+          const user = localStorage.getItem('username');
+          if (user !== channelCreator) {
+            return false;
+          } else {
+            try{
+              const response = await api.delete(`/protected/channel/${id}/user/${user.id}`, params, {
+                headers: {
+                  'Authorization': `Bearer ${token}`
+                }
+              })
+              this.channels = this.channels.map(channel => {
+                if(channel.id === id) {
+                  return response.data
+                } else {
+                  return channel
+                }
+              })
+              return response.data
+            }
+            catch(error) {
+              throw error
+            }
           }
         },
 }
